feat: move formatActivityTimestamp
