+++ ui/foundation/react/index.ts
@@ -47,7 +47,7 @@ export {
   type CreateContextReturn,
 } from "./utils/context.js";
 
-export { formatDate } from "./utils/dates.js";
+export { formatDate, formatActivityTimestamp } from "./utils/dates.js";
 export { twMerge } from "./utils/twMerge.js";
 export { mergeRefs } from "./utils/mergeRefs.js";
 export { forwardRefPolymorphic } from "./utils/polymorph.js";
