+++ ui/foundation/react/utils/dates.ts
@@ -1,4 +1,10 @@
-import { format } from "date-fns";
+import {
+  differenceInDays,
+  differenceInHours,
+  differenceInMinutes,
+  format,
+  isToday,
+} from "date-fns";
 import { formatInTimeZone } from "date-fns-tz";
 
 import type { DateArg, FormatOptions } from "date-fns";
@@ -24,3 +30,29 @@ export function formatDate(
 
   return format(date, mask, formatOptions);
 }
+
+export const formatActivityTimestamp = (timestamp: Date, mask: string): string => {
+  const now = new Date();
+  if (isToday(timestamp)) {
+    const minutesDifference = differenceInMinutes(now, timestamp);
+    if (minutesDifference < 1) {
+      return "1m";
+    }
+
+    if (minutesDifference < 60) {
+      return `${minutesDifference}m`;
+    }
+
+    const hoursDifference = differenceInHours(now, timestamp);
+    if (hoursDifference < 24) {
+      return `${hoursDifference}h`;
+    }
+  }
+
+  const daysDifference = differenceInDays(now, timestamp);
+  if (daysDifference === 1) {
+    return "1d";
+  }
+
+  return formatDate(timestamp, mask);
+};
