+++ ui/portal/web/src/components/activities/Activity.tsx
@@ -10,38 +10,10 @@ import {
 } from "react";
 import { useActivities } from "@left-curve/store";
 
-import { differenceInDays, differenceInHours, differenceInMinutes, isToday } from "date-fns";
-
-import { formatDate, IconClose, useApp } from "@left-curve/applets-kit";
+import { formatActivityTimestamp, IconClose, useApp } from "@left-curve/applets-kit";
 
 import type React from "react";
 import type { Activities, ActivityRecord } from "@left-curve/store";
-
-const formatActivityTimestamp = (timestamp: Date, mask: string): string => {
-  const now = new Date();
-  if (isToday(timestamp)) {
-    const minutesDifference = differenceInMinutes(now, timestamp);
-    if (minutesDifference < 1) {
-      return "1m";
-    }
-
-    if (minutesDifference < 60) {
-      return `${minutesDifference}m`;
-    }
-
-    const hoursDifference = differenceInHours(now, timestamp);
-    if (hoursDifference < 24) {
-      return `${hoursDifference}h`;
-    }
-  }
-
-  const daysDifference = differenceInDays(now, timestamp);
-  if (daysDifference === 1) {
-    return "1d";
-  }
-
-  return formatDate(timestamp, mask);
-};
 
 const activities: Record<
   keyof Activities,
